requirements.txt
fastapi[standard]
pydantic-settings
python-dotenv
loguru
pyjwt~=2.10.1
